rename internal chips to internal
rename
